Fixed upload for the new Filesystems
This was a bit of a tough one. The solution I found I know that is not optimal however I was struggling to find why the following code on the pages controller:

  $location = Storage::disk($disk)->getDriver()
                                            ->getAdapter()
                                            ->getPathPrefix();

was failing to work if the $disk variable was one of the "new" filesystems.

Resetting the config did not fix the issue (php artisan config:clear)

I wonder if it was because a few symlinks need to be implemented to get the new filesystems to work.
Newer versions of laravel do this automatically (php artisan storage:link
) but not 5.2.

My solution was to store the files in /files/... because this folder is already created and the symlinks are also already configured. This seemed to fix the issue.
Also removed the troublesome code.